some fixes to create project already uploaded

+++ python_web_frame/controllers/model_controller.py
@@ -23,6 +23,17 @@ class ModelController:
             cls._instance = super(ModelController, cls).__new__(cls, *args, **kwargs)
         return cls._instance
 
+    def get_already_uploaded_models(self, user):
+        already_uploaded_models = []
+        not_created_models = Dynamo().query_user_models_from_state(user, "not_created")
+        if not_created_models:
+            for model in not_created_models:
+                if not model["model_share_link_qrcode"]:
+                    self.delete_model(model, user)
+                else:
+                    already_uploaded_models.append(model)
+        return already_uploaded_models
+
     def search_models_by_name(self, search_input, user, shared=False):
 
         matching_name_models = []
@@ -236,7 +247,8 @@ class ModelController:
 
         model = self.change_model_state(model, model["model_state"], "deleted")
         S3().copy_folder_from_one_bucket_to_another(lambda_constants["processed_bucket"], lambda_constants["archive_bucket"], model["model_upload_path"], model["model_upload_path"])
-        S3().delete_folder(lambda_constants["processed_bucket"], model["model_upload_path"])
+        if model["model_state"] == "completed":
+            S3().delete_folder(lambda_constants["processed_bucket"], model["model_upload_path"])
         Dynamo().put_entity(model)
 
     def generate_new_model(self, email, filename="", federated=False, federated_required_ids=[]):
